mv silly main to test class
