+++ utils/src/test/java/com/cloud/utils/net/MacAddressTest.java
@@ -21,6 +21,7 @@ package com.cloud.utils.net;
 
 import static org.junit.Assert.assertEquals;
 
+import com.cloud.utils.*;
 import org.junit.Test;
 
 public class MacAddressTest {
@@ -57,4 +58,10 @@ public class MacAddressTest {
     // TODO    public final void testParseShort() throws Exception {
     // TODO    public final void testParseByte() throws Exception {
 
+    public static void main(String[] args) {
+        MacAddress addr = MacAddress.getMacAddress();
+        System.out.println("addr in integer is " + addr.toLong());
+        System.out.println("addr in bytes is " + NumbersUtil.bytesToString(addr.toByteArray(), 0, addr.toByteArray().length));
+        System.out.println("addr in char is " + addr.toString(":"));
+    }
 }
